Improved layout based on review feedback
